Fixed bug in gridlist splitting functionality
Previously this was generating empty gridlist files when NPROCESS is
large relative to the number of gridcells. We need to use a round-robin
distribution of gridcells to ensure that each process simulates as
close to an equal number of gridcells as possible.

Turns out that split also actually has a simpler way of doing this, so
we don't need to implement the arithmetic ourselves.