Fixed gridlist chunking for large NPROCESS
GNU split uses a default suffix length of 2, which will be exchausted
for large NPROCESS (ie >99). My solution is to compute the required
suffix length. The other option would be to just use a really large
suffix (50 should do the trick, I would think. Unless you're running
with >10^50 CPUs, in which case certain parts of this script are going
to be extremely slow anyway).