Fixed a bug in create-benchmarks script
nyear_spinup_data is actually the number of years' data in the input
file to be used during the spinup period. Previously, the script was
setting it to the total number of spinup years. It will now be set to
the total number of full years of data in the input file.