Fixed handling of date column in lai data
This column should ideally be coerced to datetime.datetime (why would
you name the class and the module the same...). Ideally, one would use
the `parse_dates = True` option to `read_csv()` and `read_excel()`.
Unfortunately, this option is only added in pands >= 2.0.0, which is
currently unavailable on my system, and likely others' systems too.
Additionally, the current method of passing in a date parser will be
removed in 2.0.0, so relying on that also seems like a recipe for
disaster. Therefore we manually call `to_datetime()` on the column.